+++ packages/hooks/useViewTransition.ts
@@ -1,5 +1,5 @@
 import { useData } from "vitepress";
-import { nextTick, provide, unref } from "vue";
+import { nextTick, provide } from "vue";
 
 /**
  * 使用 View Transition API
@@ -7,8 +7,7 @@ import { nextTick, provide, unref } from "vue";
 export const useViewTransition = (duration = 300, easing = "ease-in") => {
   const { isDark, theme } = useData();
 
-  // TODO 从 useTeekConfig 获取配置
-  const isOpenViewTransition = unref(theme).viewTransition ?? true;
+  const isOpenViewTransition = theme.value.viewTransition ?? true;
 
   if (!isOpenViewTransition) return;